Guest reservation check (user deletion)

+++ AccomodationBookingApp/Backend/user_profile_service/communication/handler/DeleteUserProfileHandler.go
@@ -1,27 +1,32 @@
 package handler
 
 import (
+	reservation "common/proto/reservation_service/generated"
 	events "common/saga/delete_user"
 	"common/saga/messaging"
+	"context"
 	"go.mongodb.org/mongo-driver/bson/primitive"
+	"user_profile_service/communication"
 	"user_profile_service/domain/model"
 	"user_profile_service/domain/service"
 	"user_profile_service/event_sourcing"
 )
 
 type DeleteUserProfileHandler struct {
-	userProfileService *service.UserProfileService
-	eventService       *event_sourcing.EventService
-	replyPublisher     messaging.Publisher
-	commandSubscriber  messaging.Subscriber
+	userProfileService        *service.UserProfileService
+	reservationServiceAddress string
+	eventService              *event_sourcing.EventService
+	replyPublisher            messaging.Publisher
+	commandSubscriber         messaging.Subscriber
 }
 
-func NewDeleteUserProfileHandler(userProfileService *service.UserProfileService, eventService *event_sourcing.EventService, replyPublisher messaging.Publisher, commandSubscriber messaging.Subscriber) (*DeleteUserProfileHandler, error) {
+func NewDeleteUserProfileHandler(userProfileService *service.UserProfileService, reservationServiceAddress string, eventService *event_sourcing.EventService, replyPublisher messaging.Publisher, commandSubscriber messaging.Subscriber) (*DeleteUserProfileHandler, error) {
 	handler := &DeleteUserProfileHandler{
-		userProfileService: userProfileService,
-		eventService:       eventService,
-		replyPublisher:     replyPublisher,
-		commandSubscriber:  commandSubscriber}
+		userProfileService:        userProfileService,
+		reservationServiceAddress: reservationServiceAddress,
+		eventService:              eventService,
+		replyPublisher:            replyPublisher,
+		commandSubscriber:         commandSubscriber}
 
 	err := handler.commandSubscriber.Subscribe(handler.handle)
 	if err != nil {
@@ -40,14 +45,33 @@ func (handler *DeleteUserProfileHandler) handle(command *events.DeleteUserComman
 
 	switch command.Type {
 	case events.DeleteUserProfile:
+		//TODO check role and reservations
+		reservationClient := communication.NewReservationClient(handler.reservationServiceAddress)
+		//TODO context namesti
+		result, err := reservationClient.GuestHasActiveReservations(context.TODO(), &reservation.GuestHasActiveReservationsRequest{GuestId: command.UserProfileId.String()})
+		if err != nil {
+
+			reply.Type = events.UserProfileDeletionFailed
+			reply.ErrorMessage = "Guest has active reservations"
+			break
+		}
+
+		if result.HasActiveReservations {
+			reply.Type = events.UserProfileDeletionFailed
+			reply.ErrorMessage = "Guest has active reservations"
+			break
+		}
+		//***********************
 		userProfileBackup, err := handler.userProfileService.GetById(command.UserProfileId)
 		if err != nil {
 			reply.Type = events.UserProfileDeletionFailed
+			break
 		}
 
 		err = handler.userProfileService.Delete(command.UserProfileId)
 		if err != nil {
 			reply.Type = events.UserProfileDeletionFailed
+			break
 		}
 
 		handler.eventService.Save(&event_sourcing.Event{
